move color helper methods into their own file
