+++ src/ui/qt/util/ColorHelpers.h
@@ -1,0 +1,18 @@
+/*
+* VGMTrans (c) 2002-2023
+* Licensed under the zlib license,
+* refer to the included LICENSE.txt file
+ */
+
+#pragma once
+
+#include <QColor>
+#include <QPalette>
+#include <QString>
+
+QString cssColor(const QColor &color);
+QColor blendColors(const QColor &foreground, const QColor &background, qreal foregroundWeight);
+QColor contrastingTextColor(const QColor &foreground, const QColor &background,
+                            const QPalette &palette,
+                            QPalette::ColorGroup colorGroup = QPalette::Active);
+bool isDarkPalette(const QPalette &palette);
